move tree and its models to their own package
