+++ src/main/java/hu/aestallon/storageexplorer/ui/tree/model/StorageInstance.java
@@ -13,7 +13,7 @@
  * If not, see <http://www.gnu.org/licenses/>.
  */
 
-package hu.aestallon.storageexplorer.model.tree;
+package hu.aestallon.storageexplorer.ui.tree.model;
 
 import java.util.ArrayList;
 import java.util.Comparator;
