+++ src/main/java/hu/aestallon/storageexplorer/ui/tree/model/StorageMap.java
@@ -13,20 +13,37 @@
  * If not, see <http://www.gnu.org/licenses/>.
  */
 
-package hu.aestallon.storageexplorer.model.tree;
+package hu.aestallon.storageexplorer.ui.tree.model;
 
-import java.util.List;
-import java.util.stream.Collectors;
 import javax.swing.tree.DefaultMutableTreeNode;
-import hu.aestallon.storageexplorer.domain.storage.model.ObjectEntry;
+import hu.aestallon.storageexplorer.domain.storage.model.MapEntry;
+import hu.aestallon.storageexplorer.domain.storage.model.StorageEntry;
 
-public final class StorageSchema extends DefaultMutableTreeNode {
+public class StorageMap extends DefaultMutableTreeNode implements Clickable {
 
-  StorageSchema(String name, List<ObjectEntry> objectEntries) {
-    super(name, true);
-    objectEntries.stream()
-        .collect(Collectors.groupingBy(ObjectEntry::typeName))
-        .forEach((type, entries) -> add(new StorageType(type, entries)));
+  public StorageMap(MapEntry mapEntry) {
+    super(mapEntry, false);
+  }
+
+  @Override
+  public StorageEntry storageEntry() {
+    return (StorageEntry) userObject;
+  }
+
+  @Override
+  public boolean getAllowsChildren() {
+    return false;
+  }
+
+  @Override
+  public boolean isLeaf() {
+    return true;
+  }
+
+  @Override
+  public String toString() {
+    final var mapEntry = (MapEntry) userObject;
+    return mapEntry.schema() + " / " + mapEntry.name();
   }
 
 }
