+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
 # Declare storage location here:
-# fs.base.directory=./fs
+fs.base.directory=./p9/dev-fs
 
 # Define how many outbound edges to traverse in a single rendering routine. 0 means no outgoing
 # edges shall be detected. Negative values mean there is no limit applied (referenced nodes are
